Completed face registration
Completed the steps needed to register a face so that the eyes are at
standard positions in the “face chip” sent to the server. The chip is
currently coded to be 150 rows by 130 columns, and the eyes are mapped
to (30,45) and (100,45), where the origin is at the top left. The steps
of the process are implemented as a series of primitive transformations
stacked together. I find Apple’s conventions to be rather confusing,
but figured it out.