Flynn - Enroll segue initial work
Added a “Finish” button to the enrolled. It doesn’t work yet.  Need to
put the face chips in a mutable array wen they’re extracted so we can
enroll them (but be careful to handle the case where we get an image
and a chip, and then operator decides to replace that image with
another one. hmm.). Assed iToast  to the project so we can have
notifications.  When enrolling, should place the enrollment processing
on a background thread because it could be up to 8 image transfers and
that might take a while.