manage and enroll all 8 images
added code to manage the (up to) 8 image acquired. Also added a set of
image buttons pushed so ew can reset the images. Added a “reset”
button, and a “go back” button that goes back to the start screen, duh.
Also added an “Identify” button to the start screen but it doesn’t do
anything yet.