10-20-15 C[4] meeting: transmit uncropped images for use with br
We are retiring eigenfaces in favor of OpenBR (openbiometrics.org).
OpenBR wants to use its own face extractor so we can’t send it cropped
chips. The app still verifies that one face is present, but transmits
the entire image to the server, not a chip.  Need to make similar
changes to the match function, on both client and server. PJF