10-27-15 added a seque to the storyboard
You can segue back to home from the identification result view
controller. Server was modified to return firstname lastname in
matching. need to unify face detection processing in enroll and id
because they are not handled the same way. Todo: make the enrollment
controller place a cropped face rectang;e in the image buttons, not the
entire image. Also added recognition to the match route in the Python
server. Need to add it to the verify route also.