12-8-15 recrop faces and standardize size to 512x512
Steve and I determined that it’s a bad thing to send the entire image
instead of a cropped face. The entire image could be big. This update
takes the face rectangle detected by the CIdetector, expands it by 25%
on each of the four sizes, crops it to the original image bounds
rectangle, and scales the image so that the longest size is 512 pixels.
The result is sent to the server.